+++ src/cosmosdb-preview/HISTORY.rst
@@ -4,8 +4,7 @@ Release History
 ===============
 0.23.0
 * Create and manage mongo clusters.
-* Add 'source_backup_location' param to 'cosmosdb restore' command
-* Add 'enable_public_network' param to 'cosmosdb restore' command
+* Add 'source_backup_location' parameter to 'cosmosdb restore' command
 
 ++++++
 0.22.0